# feat (encoder): add encoder option `NoNullSliceOrMap`