--- a/.github/workflows/license-check.yml
+++ b/.github/workflows/license-check.yml
@@ -4,11 +4,11 @@ on: push
 
 jobs:
   build:
-    runs-on: self-hosted
+    runs-on: [self-hosted, X64]
     steps:
       - uses: actions/checkout@v2
 
       - name: Check License Header
-        uses: apache/skywalking-eyes@main
+        uses: apache/skywalking-eyes/header@main
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}