--- a/encoder/encoder.go
+++ b/encoder/encoder.go
@@ -37,6 +37,7 @@ const (
     bitEscapeHTML          
     bitCompactMarshaler
     bitNoQuoteTextMarshaler
+    bitNoNullSliceOrMap
 )
 
 const (
@@ -52,11 +53,13 @@ const (
 
     // CompactMarshaler indicates that the output JSON from json.Marshaler 
     // is always compact and needs no validation 
-    CompactMarshaler   Options = 1 << bitCompactMarshaler
+    CompactMarshaler     Options = 1 << bitCompactMarshaler
 
     // NoQuoteTextMarshaler indicates that the output text from encoding.TextMarshaler 
     // is always escaped string and needs no quoting
     NoQuoteTextMarshaler Options = 1 << bitNoQuoteTextMarshaler
+
+    NoNullSliceOrMap     Options = 1 << bitNoNullSliceOrMap
 )
 
 // Encoder represents a specific set of encoder configurations.